Create a new LogSource when changing serverless tags
The "channel" LogSource is only used by the serverless agent, but could
be useful for other purposes, so this also abstracts the
serverless-specific bits (name, source) out.

More importantly, when the tags are updated, this now drops the old
source and creates a new source.  This avoids race conditions with
modifying a slice pointer while it might be in use by another gorotuine.

Since the tailers for both sources are reading from the same channel, no
messages will be lost -- a given message will either be handled by the
old tailer or, if it stops before the message is handled, will wait for
the new tailer to start.