Protect channel tailer tags with mutex
The channel tailer closes its input channel, and the channel scheduler
does not support removing sources, so this implements the async update
of the tags applied to the channel with a new field (ChannelTags)
protected by a new mutex (ChannelTagsMutex).  The new field avoids the
requirement to guard _every_ access to LogConfig.Tags with a mutex.